requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.32.1
pillow==11.0.0
python-docx==1.1.2
pypdf==6.1.3
reportlab==4.2.5
docx2pdf==0.1.8
pdf2image==1.17.0
pytesseract==0.3.13
python-multipart==0.0.20
slowapi==0.1.9
pip-audit==2.9.0
bandit==1.9.1
pdfminer.six
python-dotenv==1.2.1
starlette==0.49.1
zipfile36